Fix Rational8 overflow by upgrading to i16 storage
The Rational8 type now uses i16 internally instead of i8, which:
- Eliminates overflow risk for any realistic physics calculations
- Allows denominators up to 32,767 (vs 127 with i8)
- Uses i32 for intermediate arithmetic to prevent overflow
- Increases Dimension struct size from 22 to 44 bytes (acceptable trade-off)

Added tests verifying operations that would have overflowed with i8:
- Large denominator addition (1/12 + 1/12)
- Large numerator multiplication (50 * 50 = 2500)
- Complex fractions (1/15 * 1/15 = 1/225)
- High dimension powers (L^100)

+++ src/parsing.rs
@@ -1100,10 +1100,10 @@ fn parse_power(s: &str) -> UnitResult<Rational8> {
         let (num_str, den_str) = s.split_at(idx);
         let den_str = &den_str[1..];
 
-        let num: i8 = num_str.trim().parse().map_err(|_| {
+        let num: i16 = num_str.trim().parse().map_err(|_| {
             UnitError::ParseError(format!("invalid power numerator: {}", num_str))
         })?;
-        let den: i8 = den_str.trim().parse().map_err(|_| {
+        let den: i16 = den_str.trim().parse().map_err(|_| {
             UnitError::ParseError(format!("invalid power denominator: {}", den_str))
         })?;
 
@@ -1114,7 +1114,7 @@ fn parse_power(s: &str) -> UnitResult<Rational8> {
         Ok(Rational8::new(num, den))
     } else {
         // Simple integer power
-        let exp: i8 = s.parse().map_err(|_| {
+        let exp: i16 = s.parse().map_err(|_| {
             UnitError::ParseError(format!("invalid power: {}", s))
         })?;
         Ok(Rational8::new(exp, 1))
